fixed ProfileField
