+++ client/src/pages/ProfilePage/ProfilePage.jsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useState } from 'react'
+import React, { Fragment, useEffect, useState } from 'react'
 import ProfileCover from '../../components/ProfileCover/ProfileCover'
 import ShortUserData from '../../components/ShortUserData/ShortUserData'
 import ProfileField from '../../components/ProfileField/ProfileField'
@@ -10,11 +10,18 @@ import PostFeed from '../../components/PostFeed/PostFeed'
 
 import { Grid, Paper } from '@material-ui/core'
 import useStyles from './profilePageStyles'
+import { getUserPhotosFromPosts } from '../../actions/image'
+import { connect } from 'react-redux'
+import PropTypes from 'prop-types'
 
-const ProfilePage = () => {
+const ProfilePage = ({ user, loadUserPhotos, userPhotos, loadingPhotos }) => {
   const classes = useStyles()
-
   const [profileTab, setProfileTab] = useState('your story')
+  const { friends } = user
+
+  useEffect(() => {
+    loadUserPhotos()
+  }, [loadUserPhotos])
 
   const handleChangeTab = (event, newValue) => {
     setProfileTab(newValue)
@@ -35,10 +42,10 @@ const ProfilePage = () => {
                 <ShortUserData />
               </Paper>
               <Paper className={classes.paper}>
-                <ProfileField fieldName='Photos' />
+                <ProfileField userPhotos={userPhotos} loadingPhotos={loadingPhotos}/>
               </Paper>
               <Paper className={classes.paper}>
-                <ProfileField fieldName='Friends' />
+                <ProfileField friends={friends} />
               </Paper>
             </Grid>
             <Grid item xs={9} sm={5}>
@@ -77,4 +84,21 @@ const ProfilePage = () => {
   )
 }
 
-export default ProfilePage
+ProfilePage.propTypes = {
+  user: PropTypes.object.isRequired,
+  loadUserPhotos: PropTypes.func.isRequired,
+  userPhotos: PropTypes.array.isRequired,
+  loadingPhotos: PropTypes.bool.isRequired
+}
+
+const mapStateToProps = state => ({
+  user: state.auth.user,
+  userPhotos: state.images.userPhotos,
+  loadingPhotos: state.images.loading
+})
+
+const mapDispatchToProps = dispatch => ({
+  loadUserPhotos: () => dispatch(getUserPhotosFromPosts())
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(ProfilePage)
